refactor(server): group by domain and enforce layering in lint
Server code was organised by layer (routers/, services/, repositories/),
so a change to one domain meant editing three sibling directories. Nobody
edits along the layer axis; every real change is domain-shaped. This moves
the six domains into one folder each:

  server/<domain>/{router,service,repository}.ts (+ service.spec.ts)

server/api/root.ts keeps its role as the registry, and api/trpc.ts still
holds context and the procedure builders.

Note this also makes the server consistent with the frontend, which the
tRPC migration already organised feature-first under features/<name>/.

The layering itself is now enforced rather than documented: biome.json
gains two noRestrictedImports overrides at error severity, so a router
importing a repository, or a repository importing a service or router,
fails `bun run lint`. Verified by injecting both violations and confirming
they are caught. This restores the compile-time boundary the NestJS module
graph used to provide (ADR 0002).

No behaviour change: pure file moves plus import rewrites. Services and
repositories kept their directory depth, so ../db, ../errors and ../ai
resolve unchanged; only routers moved depth (../trpc -> ../api/trpc).
Cross-domain service calls are preserved (search/service.ts still imports
../course/service).

CLAUDE.md and AGENTS.md documented the old layout and have been updated.
