--- apps/web/server/api/routers/health.ts
@@ -1,6 +1,0 @@
-import { testAll } from "@/server/services/health";
-import { baseProcedure, createTRPCRouter } from "../trpc";
-
-export const healthRouter = createTRPCRouter({
-  check: baseProcedure.query(() => testAll()),
-});
