+++ apps/web/server/health/router.ts
@@ -1,0 +1,6 @@
+import { baseProcedure, createTRPCRouter } from "../api/trpc";
+import { testAll } from "./service";
+
+export const healthRouter = createTRPCRouter({
+  check: baseProcedure.query(() => testAll()),
+});
