+++ apps/web/server/user/service.ts
@@ -1,4 +1,4 @@
-import * as userRepo from "../repositories/user";
+import * as userRepo from "./repository";
 
 export type Me = {
   userId: string;
